backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.1
numpy==2.1.3
python-dotenv==1.0.1
